python_parser/requirements.txt
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
cssutils==2.11.1
frozendict==2.4.6
idna==3.10
loguru==0.7.3
lxml==5.4.0
more-itertools==10.8.0
numpy==2.2.6
pandas==2.3.2
pyrate-limiter==3.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
sec-downloader==0.11.2
sec-edgar-downloader==5.0.3
sec-parser==0.58.1
sec-parsers==0.549
six==1.17.0
soupsieve==2.8
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0

# API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database dependencies (for comprehensive financial calculator)
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Semantic matching dependencies (Phase 3)
# Updated for Python 3.13 compatibility
sentence-transformers==3.3.1
scikit-learn==1.6.1
torch==2.6.0
